Some edits to get it to compile in Visual Studio

+++ src/demo/star_bmp.h
@@ -7,8 +7,8 @@ class StarBmp
 	//Initializes variables
 	StarBmp();
 
-	int x;
-	int y;
+	double x;
+	double y;
 	int spd;
-	int bmp;
+	double bmp;
 };
